int -> size_t in suffstat proxy template in branchlength_sm

+++ src/submodels/branchlength_sm.hpp
@@ -59,7 +59,7 @@ struct branchlengths_sm {
 
     template <class BLModel, class Gen>
     static auto gibbs_resample(
-        BLModel& model, Proxy<PoissonSuffStat&, int>& ss, Gen& gen) {
+        BLModel& model, Proxy<PoissonSuffStat&, size_t>& ss, Gen& gen) {
         /* -- */
         auto& raw_vec = get<bl_array, value>(model);
